requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
openai
mysql-connector-python  # optional if you use mysql
